Feat: feature.search 모듈 build-logic 적용
