+++ feature/main/src/main/java/com/kolown/porring/feature/main/component/MainNavHost.kt
@@ -20,7 +20,7 @@ import com.kolown.model.Reactions
 import com.kolown.model.UploadModel
 import com.kolown.porring.feature.my.navigation.myNavGraph
 import com.kolown.porring.feature.home.navigation.homeNavGraph
-import com.kolown.search.navigation.searchNavGraph
+import com.kolown.porring.feature.search.navigation.searchNavGraph
 import com.kolown.setting.navigation.settingNavGraph
 import com.kolown.their.navigation.theirNavGraph
 import com.kolown.upload.navigation.uploadNavGraph
