+++ feature/search/src/main/java/com/kolown/porring/feature/search/DetailSearchScreen.kt
@@ -1,4 +1,4 @@
-package com.kolown.search
+package com.kolown.porring.feature.search
 
 
 import androidx.activity.compose.BackHandler
