+++ feature/search/src/main/java/com/kolown/porring/feature/search/component/PostHeader.kt
@@ -1,6 +1,5 @@
-package com.kolown.search.component
+package com.kolown.porring.feature.search.component
 
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
@@ -9,21 +8,18 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
 import com.kolown.model.Tag
-import com.kolown.search.R
 
 @Composable
-internal fun TagItem(tag: Tag, onClick: (Tag) -> Unit) {
+internal fun PostHeader(tag: Tag) {
     Row(
         modifier = Modifier
             .fillMaxWidth()
-            .height(48.dp)
-            .padding(horizontal = 16.dp)
-            .clickable { onClick(tag) },
+            .height(48.dp),
         verticalAlignment = Alignment.CenterVertically
     ) {
-        Text(stringResource(R.string.string_tag_name, tag.name))
+        Text(fontSize = 22.sp, text = "# ${tag.name}")
     }
 }
