+++ feature/search/src/main/java/com/kolown/porring/feature/search/component/SearchResult.kt
@@ -1,13 +1,9 @@
-package com.kolown.search.component
+package com.kolown.porring.feature.search.component
 
-import android.util.Log
 import androidx.compose.animation.animateContentSize
-import androidx.compose.foundation.Canvas
 import androidx.compose.foundation.background
-import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.lazy.LazyColumn
@@ -17,13 +13,12 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.paging.compose.LazyPagingItems
 import com.kolown.designsystem.ui.theme.Gray
 import com.kolown.model.Tag
-import com.kolown.search.R
+import com.kolown.porring.feature.search.R
 
 @Composable
 fun SearchResult(
