+++ feature/search/src/main/java/com/kolown/porring/feature/search/component/TagItem.kt
@@ -1,5 +1,6 @@
-package com.kolown.search.component
+package com.kolown.porring.feature.search.component
 
+import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
@@ -8,18 +9,21 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
 import com.kolown.model.Tag
+import com.kolown.porring.feature.search.R
 
 @Composable
-internal fun PostHeader(tag: Tag) {
+internal fun TagItem(tag: Tag, onClick: (Tag) -> Unit) {
     Row(
         modifier = Modifier
             .fillMaxWidth()
-            .height(48.dp),
+            .height(48.dp)
+            .padding(horizontal = 16.dp)
+            .clickable { onClick(tag) },
         verticalAlignment = Alignment.CenterVertically
     ) {
-        Text(fontSize = 22.sp, text = "# ${tag.name}")
+        Text(stringResource(R.string.string_tag_name, tag.name))
     }
 }
